Flash-Lang-main-RevA/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
youtube-transcript-api>=0.6.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.0
psycopg2-binary>=2.9.0
